+++ contracts/test/Dependencies.sol
@@ -1,9 +1,0 @@
-pragma solidity 0.5.0;
-
-// hack for truffle, so we have artifact for dependencies contracts for migrations and tests
-
-import "contract-registry/contracts/ContractRegistry.sol";
-import "token-incentivized-sidechains/contracts/StakingBank.sol";
-import "token-sale-contracts/contracts/HumanStandardToken.sol";
-import "digivice/contracts/VerifierRegistry.sol";
-import "digivice/contracts/VerifierRegistryStorage.sol";